calculation changed for year table

+++ pathfinder_be/src/main/java/com/example/pathfinder/dto/ItRunSpendDto.java
@@ -17,7 +17,7 @@ public class ItRunSpendDto {
 	private long baseLineYearSpend;
 	
 	private long total_itSpendWithInflation;
-    private double total_cotClientShare;
+    private long total_cotClientShare;
     private long total_yearlyRunRate;
 	
 	private List<ItRunYearlyDto> runCalc;
