+++ pathfinder_be/src/main/java/com/example/pathfinder/dto/ItSpendOnRunPersonnelYear.java
@@ -6,8 +6,8 @@ import lombok.Data;
 
 public class ItSpendOnRunPersonnelYear {
 	private int year;
-	private double itSpendWithInflation;
-	private double cotClientShare;
-	private double yearlyRunRate;
+	private long itSpendWithInflation;
+	private long cotClientShare;
+	private long yearlyRunRate;
 	
 }
